Indenting, multiple channels and a channel object, delegate method grammar
Changed indenting to 4 spaces per level as per Objective-C standards
Added the ability to subscribe to (and for the FayeClient to track) multiple channels using the same FayeClient instance.
Added a FayeChannel object for easier encapsulation of any metadata associated with a specific channel.
Fixed delegate method grammar as per Objective-C/Cocoa standards.
Only sends up websocket as a supported connection type, since that *is* the only supported connection type!